requirements.txt
# Core dependencies
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-google-genai>=0.0.6
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0
langchain-experimental>=0.0.29

# Database
sqlalchemy>=2.0.0

# API
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
httpx>=0.25.2

# Speech
azure-cognitiveservices-speech>=1.30.0

# Data science
numpy>=1.24.0
scikit-learn>=1.2.0

# PDF Generation
reportlab>=4.0.0

# Google API
google-generativeai>=0.3.0
google-api-python-client>=2.108.0
google-auth>=2.23.4
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Utilities
pydantic>=2.5.0 